changed respondant partner id to long

+++ src/main/java/com/employmeo/data/model/Respondant.java
@@ -82,7 +82,7 @@ public class Respondant implements Serializable {
 	private Partner partner;
 
 	@Column(name = "respondant_partner_id", insertable = true, updatable = true)
-	private Integer partnerId;
+	private Long partnerId;
 
 	@JsonBackReference
 	@ManyToOne
